fixed bugs (but not menu)

+++ DivisionKuk.cpp
@@ -1,7 +1,7 @@
 #include "DivisionKuk.h"
 #include "InverseKuk.h"
 
-double DivisionKuk::divide(int x, int divisor) {
+double DivisionKuk::divide(int x, double divisor) {
     InverseKuk inverseKuk;
     return ((double) x) * inverseKuk.getInverse(divisor);
 }
